requirements.txt
ConfigSpace==0.4.12
cycler==0.10.0
Cython==0.29.17
docutils==0.16
joblib==0.14.1
kiwisolver==1.1.0
lazy-import==0.2.2
matplotlib==3.1.2
mpmath==1.1.0
numpy==1.18.1
psutil==5.7.0
pynisher==0.5.0
pyparsing==2.4.6
pyrfr==0.8.0
python-dateutil==2.8.1
scikit-learn==0.21.3
scipy==1.4.1
six==1.13.0
smac==0.12.0
sobol-seq==0.1.2
sympy==1.5.1
jupyter
PyYAML
plotly
progressbar
pympler
scikit-fmm
attrdict
pandas==0.25.1
seaborn==0.9.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torch==1.4.0
pyquaternion==0.9.5
pytest==5.3.0
orjson==2.1.4
ncls==0.0.51
dill==0.3.1.1
tqdm==4.45.0
notebook==6.0.3
opencv-python==4.1.1.26
nuscenes-devkit==1.0.6
cvxpy == 1.1.13